+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG
 =========
 
-2.4, 2024-11-12
+2.4, 2024-11-14
 ---------------
 - add support for generated columns
 - add find/replace dialog to data grids and column editors and database schema text
@@ -26,6 +26,7 @@ CHANGELOG
 - use hourglass cursor while showing busy message
 - allow saving SVG image as SVG in column editor
 - provide workaround for importing CSVs with mixed Windows and Unix linefeeds (issue #3)
+- fix data grids lagging with very long column values in Linux (issue #4)
 - fix single instance checker stalling if existing instance has hung
 - fix error on cancelling export to another database
 - fix rolling back changes on error in schema editor test and save